Attempt at implementation of size_t aware MPI behaviour. Possibly quite broken at this stage.
git-svn-id: svn://thep.housing.rug.nl/lemon@40 038c52d2-56c5-4035-8d8a-1cd00843ba61

+++ include/lemon_reader.h
@@ -50,7 +50,7 @@ typedef struct
 
   /* Data needed for tracking I/O requests */
   void *buffer;
-  int bytes_wanted;
+  size_t bytes_wanted;
 } LemonReader;
 
 /* Reader manipulators */
